Improve hashing performance
