+++ c/src/CMakeLists.txt
@@ -8,7 +8,7 @@ set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/lib)
 link_directories( ${CMAKE_CURRENT_BINARY_DIR} )
 include_directories (${CMAKE_CURRENT_BINARY_DIR} ${PROJECT_BINARY_DIR}/src )
 
-add_library (variantkey astring.c astring.h binsearch.c binsearch.h esid.c esid.h genoref.c genoref.h nrvk.c nrvk.h regionkey.c regionkey.h rsidvar.c rsidvar.h variantkey.c variantkey.h)
+add_library (variantkey binsearch.c binsearch.h esid.c esid.h genoref.c genoref.h hex.c hex.h nrvk.c nrvk.h regionkey.c regionkey.h rsidvar.c rsidvar.h variantkey.c variantkey.h)
 target_include_directories (variantkey PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 # Required to link the math library
