+++ c/test/CMakeLists.txt
@@ -26,11 +26,11 @@ endfunction(SMOKE_TEST)
 file(GLOB TEST_BIN_FILES "data/*.bin")
 file (COPY ${TEST_BIN_FILES} DESTINATION ${CMAKE_CURRENT_BINARY_DIR})
 
-SMOKE_TEST (test_astring test_astring.c variantkey)
 SMOKE_TEST (test_binsearch test_binsearch.c variantkey)
 SMOKE_TEST (test_esid test_esid.c variantkey)
 SMOKE_TEST (test_example test_example.c variantkey)
 SMOKE_TEST (test_genoref test_genoref.c variantkey)
+SMOKE_TEST (test_hex test_hex.c variantkey)
 SMOKE_TEST (test_nrvk test_nrvk.c variantkey)
 SMOKE_TEST (test_regionkey test_regionkey.c variantkey)
 SMOKE_TEST (test_test_rsidvar test_rsidvar.c variantkey)
