+++ c/test/test_genoref.c
@@ -54,6 +54,82 @@ uint64_t get_time()
     return (((uint64_t)t.tv_sec * 1000000000) + (uint64_t)t.tv_nsec);
 }
 
+int test_aztoupper()
+{
+    int errors = 0;
+    int i, c;
+    for (i=97 ; i <= 122; i++)
+    {
+        c = aztoupper(i);
+        if (c != (i - 32))
+        {
+            fprintf(stderr, "%s : Wrong uppercase value for %d - expecting %d, got %d\n", __func__, i, (i - 32), c);
+            ++errors;
+        }
+    }
+    c = aztoupper(96);
+    if (c != 96)
+    {
+        fprintf(stderr, "%s : Wrong uppercase value - expecting 96, got %d\n", __func__, c);
+        ++errors;
+    }
+    return errors;
+}
+
+void benchmark_aztoupper()
+{
+    int c;
+    uint64_t tstart, tend;
+    int i, j;
+    int size = 100000;
+    tstart = get_time();
+    for (i=0 ; i < size; i++)
+    {
+        for (j=0 ; j < 256; j++)
+        {
+            c = aztoupper(j);
+        }
+    }
+    tend = get_time();
+    fprintf(stdout, " * %s : %lu ns/op (%d)\n", __func__, (tend - tstart)/(256*size), c);
+}
+
+int test_prepend_char()
+{
+    int errors = 0;
+    char original[5] =   "BCD";
+    char expected[5] = "ABCD";
+    size_t size = 3;
+    prepend_char('A', original, &size);
+    if (size != 4)
+    {
+        fprintf(stderr, "%s : Expected size 4, got %lu\n", __func__, size);
+        ++errors;
+    }
+    if (strcmp(original, expected) != 0)
+    {
+        fprintf(stderr, "%s : Expected %s, got %s\n", __func__, expected, original);
+        ++errors;
+    }
+    return errors;
+}
+
+void benchmark_prepend_char()
+{
+    char original[1002] =   "B";
+    size_t len = 1;
+    uint64_t tstart, tend;
+    int i;
+    int size = 1000;
+    tstart = get_time();
+    for (i=0 ; i < size; i++)
+    {
+        prepend_char('A', original, &len);
+    }
+    tend = get_time();
+    fprintf(stdout, " * %s : %lu ns/op\n", __func__, (tend - tstart)/size);
+}
+
 int test_swap_sizes()
 {
     int errors = 0;
@@ -320,6 +396,9 @@ int main()
         fprintf(stderr, "Expecting size %" PRIu64 ", got instead: %" PRIu32 "\n", genoref.size, idx[26]);
         return 1;
     }
+
+    errors += test_aztoupper();
+    errors += test_prepend_char();
     errors += test_swap_sizes();
     errors += test_swap_alleles();
     errors += test_get_genoref_seq(genoref.src, idx);
@@ -327,6 +406,8 @@ int main()
     errors += test_flip_allele();
     errors += test_normalize_variant(genoref.src, idx);
 
+    benchmark_aztoupper();
+    benchmark_prepend_char();
     benchmark_get_genoref_seq(genoref.src, idx);
     benchmark_flip_allele();
